requirements.txt
aiohappyeyeballs==2.6.1
aiohttp==3.12.15
aiosignal==1.4.0
annotated-types==0.7.0
anyio==4.10.0
async-timeout==5.0.1
attrs==25.3.0
beautifulsoup4==4.13.4
bs4==0.0.2
certifi==2025.8.3
cffi==1.17.1
chardet==5.2.0
charset-normalizer==3.4.2
colorlog==6.9.0
cryptography==45.0.6
cssselect==1.3.0
dashscope==1.24.1
distro==1.9.0
exceptiongroup==1.3.0
fake-useragent==2.2.0
frozenlist==1.7.0
gne==0.3.1
h11==0.16.0
httpcore==1.0.9
httpx==0.28.1
idna==3.10
jiter==0.10.0
lxml==6.0.0
lxml_html_clean==0.4.2
multidict==6.6.3
numpy==2.2.6
openai==1.99.3
pip==25.1
propcache==0.3.2
pycparser==2.22
pydantic==2.11.7
pydantic_core==2.33.2
PyYAML==6.0.2
readability-lxml==0.8.4.1
regex==2025.7.34
requests==2.32.4
setuptools==78.1.1
sniffio==1.3.1
soupsieve==2.7
tavily-python==0.7.10
tiktoken==0.10.0
tqdm==4.67.1
typing_extensions==4.14.1
typing-inspection==0.4.1
urllib3==2.5.0
websocket-client==1.8.0
wheel==0.45.1
yarl==1.20.1
